+++ pkg/querier/distributor_queryable_test.go
@@ -214,6 +214,118 @@ func TestIngesterStreaming(t *testing.T) {
 	require.NoError(t, seriesSet.Err())
 }
 
+func TestIngesterStreamingMixedResults(t *testing.T) {
+	const (
+		mint = 0
+		maxt = 10000
+	)
+	s1 := []client.Sample{
+		{Value: 1, TimestampMs: 1000},
+		{Value: 2, TimestampMs: 2000},
+		{Value: 3, TimestampMs: 3000},
+		{Value: 4, TimestampMs: 4000},
+		{Value: 5, TimestampMs: 5000},
+	}
+	s2 := []client.Sample{
+		{Value: 1, TimestampMs: 1000},
+		{Value: 2.5, TimestampMs: 2500},
+		{Value: 3, TimestampMs: 3000},
+		{Value: 5.5, TimestampMs: 5500},
+	}
+
+	mergedSamplesS1S2 := []client.Sample{
+		{Value: 1, TimestampMs: 1000},
+		{Value: 2, TimestampMs: 2000},
+		{Value: 2.5, TimestampMs: 2500},
+		{Value: 3, TimestampMs: 3000},
+		{Value: 4, TimestampMs: 4000},
+		{Value: 5, TimestampMs: 5000},
+		{Value: 5.5, TimestampMs: 5500},
+	}
+
+	d := &mockDistributor{}
+	d.On("QueryStream", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(
+		&client.QueryStreamResponse{
+			Chunkseries: []client.TimeSeriesChunk{
+				{
+					Labels: []client.LabelAdapter{{Name: labels.MetricName, Value: "one"}},
+					Chunks: convertToChunks(t, s1),
+				},
+				{
+					Labels: []client.LabelAdapter{{Name: labels.MetricName, Value: "two"}},
+					Chunks: convertToChunks(t, s1),
+				},
+			},
+
+			Timeseries: []client.TimeSeries{
+				{
+					Labels:  []client.LabelAdapter{{Name: labels.MetricName, Value: "two"}},
+					Samples: s2,
+				},
+				{
+					Labels:  []client.LabelAdapter{{Name: labels.MetricName, Value: "three"}},
+					Samples: s1,
+				},
+			},
+		},
+		nil)
+
+	ctx := user.InjectOrgID(context.Background(), "0")
+	queryable := newDistributorQueryable(d, true, mergeChunks, 0)
+	querier, err := queryable.Querier(ctx, mint, maxt)
+	require.NoError(t, err)
+
+	seriesSet := querier.Select(true, &storage.SelectHints{Start: mint, End: maxt}, labels.MustNewMatcher(labels.MatchRegexp, labels.MetricName, ".*"))
+	require.NoError(t, seriesSet.Err())
+
+	require.True(t, seriesSet.Next())
+	verifySeries(t, seriesSet.At(), labels.Labels{{Name: labels.MetricName, Value: "one"}}, s1)
+
+	require.True(t, seriesSet.Next())
+	verifySeries(t, seriesSet.At(), labels.Labels{{Name: labels.MetricName, Value: "three"}}, s1)
+
+	require.True(t, seriesSet.Next())
+	verifySeries(t, seriesSet.At(), labels.Labels{{Name: labels.MetricName, Value: "two"}}, mergedSamplesS1S2)
+
+	require.False(t, seriesSet.Next())
+	require.NoError(t, seriesSet.Err())
+}
+
+func verifySeries(t *testing.T, series storage.Series, l labels.Labels, samples []client.Sample) {
+	require.Equal(t, l, series.Labels())
+
+	it := series.Iterator()
+	for _, s := range samples {
+		require.True(t, it.Next())
+		require.Nil(t, it.Err())
+		ts, v := it.At()
+		require.Equal(t, s.Value, v)
+		require.Equal(t, s.TimestampMs, ts)
+	}
+	require.False(t, it.Next())
+	require.Nil(t, it.Err())
+}
+
+func convertToChunks(t *testing.T, samples []client.Sample) []client.Chunk {
+	// We need to make sure that there is atleast one chunk present,
+	// else no series will be selected.
+	promChunk, err := encoding.NewForEncoding(encoding.Bigchunk)
+	require.NoError(t, err)
+
+	for _, s := range samples {
+		c, err := promChunk.Add(model.SamplePair{Value: model.SampleValue(s.Value), Timestamp: model.Time(s.TimestampMs)})
+		require.NoError(t, err)
+		require.Nil(t, c)
+	}
+
+	clientChunks, err := chunkcompat.ToChunks([]chunk.Chunk{
+		chunk.NewChunk("", 0, nil, promChunk, model.Time(samples[0].TimestampMs), model.Time(samples[len(samples)-1].TimestampMs)),
+	})
+	require.NoError(t, err)
+
+	return clientChunks
+}
+
 type mockDistributor struct {
 	mock.Mock
 }